pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "aiokafka>=0.13.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.134.0",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.3.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "litellm>=1.81.16",
    "mcp>=1.26.0",
    "numpy>=2.4.2",
    "openai-agents>=0.10.2",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
